requirements.txt
pyyaml>=5.3.1,<6
kubernetes>=12.0.0,<13
requests>=2.25.1,<3
jinja2>=2.11.3<3
